[refactor] supabase-config 브라우저/서버 파일 분리로 클라이언트 번들 최적화
- supabase-browser.ts: CC 전용 (createBrowserSupabaseClient)
- supabase-server.ts: 서버 전용 (createServerSupabaseClient, createServerSupabaseStorageClient)
- supabase-config.ts 삭제
