+++ nextjs/src/app/api/chat/unread-count/route.ts
@@ -1,4 +1,4 @@
-import { createServerSupabaseClient } from '@/lib/supabase-config';
+import { createServerSupabaseClient } from '@/lib/supabase-server';
 import { NextRequest, NextResponse } from 'next/server';
 import { calculateUnreadCount } from '../_utils/calculateUnreadCount';
