+++ nextjs/src/app/api/gatherings/route.ts
@@ -1,5 +1,5 @@
 import type { UpsertGatheringRequest } from '@/features/gatherings/types/gatheringTypes';
-import { createServerSupabaseClient } from '@/lib/supabase-config';
+import { createServerSupabaseClient } from '@/lib/supabase-server';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
