+++ nextjs/src/app/api/request/respond/route.ts
@@ -1,4 +1,4 @@
-import { createServerSupabaseClient } from '@/lib/supabase-config';
+import { createServerSupabaseClient } from '@/lib/supabase-server';
 
 import { NextRequest, NextResponse } from 'next/server';
