+++ nextjs/src/app/api/users/nearby-radius/route.ts
@@ -5,7 +5,7 @@ import {
   getBoundingBox,
   getExcludedUserIds,
 } from '@/app/api/_utils/location';
-import { createServerSupabaseClient } from '@/lib/supabase-config';
+import { createServerSupabaseClient } from '@/lib/supabase-server';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
