+++ nextjs/src/app/api/users/profile/route.ts
@@ -1,5 +1,5 @@
 import { fuzzyCoordinate } from '@/app/api/_utils/location';
-import { createServerSupabaseClient } from '@/lib/supabase-config';
+import { createServerSupabaseClient } from '@/lib/supabase-server';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
