+++ nextjs/src/lib/supabase-browser.ts
@@ -1,0 +1,9 @@
+import { createBrowserClient } from '@supabase/ssr';
+
+// 브라우저용 Supabase 클라이언트 (Auth 포함, 클라이언트 컴포넌트에서만 사용)
+export const createBrowserSupabaseClient = () => {
+  return createBrowserClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+  );
+};
